first commit with text selection
added the facility to highlight text in continuous and discontinuous
mode.
modified buttons in the teacher interface.
added the menu for the selection of grammar categories in the teacher
interface.
modified the "complex exercise" function in the front end and in the
backend in order to send  the grammar category from the teacher
interface to the student interface.